PMM-544 cleanup
