+++ pmm-app/src/react-plugins-deps/components/FormComponents/Checkbox/Checkbox.scss
@@ -18,6 +18,10 @@
   user-select: none;
 }
 
+.checkbox-container.no-gap {
+  grid-column-gap: 0;
+}
+
 /* Styles for main checkbox*/
 .checkbox-container--main {
   font-weight: 700;
